Move some testAccPreCheck functions to 'provider_test.go'.
