+++ aws/resource_aws_instance_test.go
@@ -4567,40 +4567,3 @@ func testAccPreCheckOffersEc2InstanceType(t *testing.T, instanceType string) {
 		t.Fatalf("error describing EC2 instance type offerings: %s", err)
 	}
 }
-
-// testAccPreCheckHasDefaultVpc checks that the test region has a default VPC.
-func testAccPreCheckHasDefaultVpc(t *testing.T) {
-	if !testAccHasDefaultVpc(t) {
-		region := testAccProvider.Meta().(*AWSClient).region
-		t.Skipf("skipping tests; %s does not have a default VPC", region)
-	}
-}
-
-// testAccPreCheckHasDefaultVpcOrEc2Classic checks that the test region has a default VPC or has the EC2-Classic platform.
-// This check is useful to ensure that an instance can be launched without specifying a subnet.
-func testAccPreCheckHasDefaultVpcOrEc2Classic(t *testing.T) {
-	client := testAccProvider.Meta().(*AWSClient)
-
-	if !testAccHasDefaultVpc(t) && !hasEc2Classic(client.supportedplatforms) {
-		t.Skipf("skipping tests; %s does not have a default VPC or EC2-Classic", client.region)
-	}
-}
-
-func testAccHasDefaultVpc(t *testing.T) bool {
-	conn := testAccProvider.Meta().(*AWSClient).ec2conn
-
-	resp, err := conn.DescribeAccountAttributes(&ec2.DescribeAccountAttributesInput{
-		AttributeNames: aws.StringSlice([]string{ec2.AccountAttributeNameDefaultVpc}),
-	})
-	if testAccPreCheckSkipError(err) ||
-		len(resp.AccountAttributes) == 0 ||
-		len(resp.AccountAttributes[0].AttributeValues) == 0 ||
-		aws.StringValue(resp.AccountAttributes[0].AttributeValues[0].AttributeValue) == "none" {
-		return false
-	}
-	if err != nil {
-		t.Fatalf("error describing EC2 account attributes: %s", err)
-	}
-
-	return true
-}
